fix(cloudflare): extract email callback to break PluginBridge chunk split
runner.ts imported setEmailSendCallback from bridge.ts, causing Rollup
to treat bridge.ts as a shared module and bundle it into a chunk.
entry.mjs then re-exported PluginBridge from the chunk, but Cloudflare's
`exports` API (cloudflare:workers) only resolves WorkerEntrypoint subclasses
that are defined inline in the entry module — chunk re-exports return null.

Fix: extract emailSendCallback state and setEmailSendCallback/getEmailSendCallback
into a new email-callback.ts module. runner.ts now imports from email-callback.ts
instead of bridge.ts, severing the cross-path dependency. bridge.ts is then
only reachable from the worker entry path, so Rollup inlines it directly into
entry.mjs, making exports.PluginBridge non-null.

Fixes #1183
